+++ RenderToMercator.h
@@ -27,4 +27,4 @@
 #include "SphericalPixelSource.h"
 
 
-FloatPixMapRef RenderToMercator(size_t size, RenderFlags flags, SphericalPixelSourceFunction source, void *sourceContext, ProgressCallbackFunction progress, ErrorCallbackFunction error, void *cbContext);
+FloatPixMapRef RenderToMercator(uintmax_t size, RenderFlags flags, SphericalPixelSourceFunction source, void *sourceContext, ProgressCallbackFunction progress, ErrorCallbackFunction error, void *cbContext);
